translator: restitch stable subset of marked safe autotranslations

+++ js/worm.js
@@ -539,7 +539,7 @@ function _rnd_nextto_goodpos(map, mtmp, x, y, player) {
 // ========================================================================
 // Autotranslated from worm.c:826
 export function size_wseg(worm) {
-  return (int) (count_wsegs(worm) * sizeof );
+  return  (count_wsegs(worm) * sizeof );
 }
 
 // ========================================================================
